feat: Add service worker to cache static assets
NOTE : This commit adds a service worker to the project in order to cache static assets. The service worker is responsible for caching the following files:
- index.html
- images/banner.png
- images/c-extension.png
- images/copilot-extension.png
- images/cs-dev-kit-extension.png
- images/github-pull-requests-extension.png
- images/gitlens-extension.png
- images/remote-extension.png
- images/jupyter-extension.png
- images/microsoft.svg
- images/python-extension.png
- images/search.svg
- images/sun.svg
- images/x-icon.svg
- images/vscode.png
- images/swimlane-anywhere.png
- images/swimlane-copilot.png
- images/swimlane-customized.png

The service worker is installed during the "install" event and the specified files are added to the cache. After the files are added, the service worker skips waiting and becomes active.